refactor(bridge): extract chat settings FFI seam
