+++ whatsrust/lib/main.go
@@ -17,13 +17,6 @@ typedef struct {
 } Contact;
 
 typedef struct {
-	bool found;
-	int64_t muted_until;
-	bool pinned;
-	bool archived;
-} ChatSettings;
-
-typedef struct {
 	char* text;
 } TextMessage;
 
@@ -618,30 +611,6 @@ func C_ForwardMessage(sourceID *C.char, sourceChat C.JID, sourceSender C.JID, so
 		destinationStrings,
 		forwardingSourceBytes(forwardSource, forwardSourceLen),
 	).cResult()
-}
-
-//export C_GetChatSettings
-func C_GetChatSettings(cjid C.JID) C.ChatSettings {
-	ctx := context.Background()
-	jid := cToJid(cjid).ToNonAD()
-	settings, err := lookupChatSettings(
-		ctx,
-		jid,
-		client.Store.ChatSettings.GetChatSettings,
-		client.Store.LIDs.GetLIDForPN,
-		client.Store.LIDs.GetPNForLID,
-	)
-	if err != nil {
-		LOG_WARN("failed to get chat settings for %s: %v", jid, err)
-		return C.ChatSettings{}
-	}
-	payload := chatSettingsPayloadFrom(settings)
-	return C.ChatSettings{
-		found:       C.bool(payload.found),
-		muted_until: C.int64_t(payload.mutedUntil),
-		pinned:      C.bool(payload.pinned),
-		archived:    C.bool(payload.archived),
-	}
 }
 
 //export C_MarkAsRead
